Revert "added files to support x86-64 codegen from instructions"
This reverts commit 328f9bb5f5dd45c09feb8c292602060d8004c59a.

+++ exp/include/env/string_interner.h
@@ -27,8 +27,8 @@
  * ptr comparison
  */
 typedef struct StringInterner {
-    u32 capacity;
-    u32 count;
+    u64 capacity;
+    u64 count;
     ConstantString **buffer;
 } StringInterner;
